PageContent service refactor
- addContentItemsToPageContent business logic refactored
- Changed request classes import packages
